requirements.txt
fastapi==0.104.1
uvicorn==0.19.0
SQLAlchemy==1.4.41
requests==2.28.1
pydantic==2.5.2
python-dotenv==1.0.1
aiohttp==3.9.1
python-dateutil==2.8.2
tzdata==2025.2